implemented websocket

+++ src/main/java/de/abstractolotl/azplace/model/requests/PlaceRequest.java
@@ -7,6 +7,6 @@ public class PlaceRequest {
 
     private int x;
     private int y;
-    private byte color;
+    private int color;
 
 }
